requirements.txt
django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt[crypto]==5.3.1
pillow==11.0.0
psycopg[binary]==3.1.10
environs==11.0.0
gunicorn==23.0.0
python-dotenv==1.0.1
whitenoise==6.7.0
django-storages==1.14.4
boto3==1.35.15
pillow=11.0.0
